Initial Commit
This version of the code is working, but there are missing features. The full feature list is as follows with X's next to the features currently not implemented:

-   Apple Intelligence integration that handles story flow.
-   Inventory and settings view
-   Story can add items to inventory
-   Story can remove items from inventory (X)
-   Story can add gold
-   Story can remove gold (X)
-   Story can injure the player (There is a temporary implementation that has little depth. We'll put X for now)
-   Story endsings handling win/loss(X)

Of little importance until all features are working properly:

-   Opening splash screen with instructions and opening text
-   Background animations and better liquid glass implementation
-   Better macOS and iOS platform switch handling

These features and bugs will be added and squashed in the future.

This version does not currently run on macOS. However, the fix is rather simple. Just need to implement different button handling in SettingsView and InventoryView.